Move request types to a different file
